feat: implement SearchBar component with conditional visibility
Overview:

Added the SearchBar component that provides a search input field with a toggleable visibility based on the route.
Used useContext to manage search state and visibility with ShopContext.
Utilized useState for local component state management and useEffect to control the visibility of the search bar based on the current route.
Integrated useLocation from react-router-dom to determine the current route and adjust the search bar’s visibility accordingly.